# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StartupCRM API",
    "description": "API para gestión de CRM - Autenticación, Emails, Contactos",
    "version": "1.0.0",
    "contact": {
      "name": "StartupCRM Support",
      "email": "support@startupcrm.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/email/receive": {
      "get": {
        "summary": "Recibir lista de emails",
        "description": "Obtiene una lista de emails desde IMAP. **Requiere autenticación**",
        "tags": ["Email"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 20
            },
            "description": "Límite de emails"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 0
            },
            "description": "Offset para paginación"
          },
          {
            "name": "unreadOnly",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Solo no leídos"
          },
          {
            "name": "mailbox",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "INBOX"
            },
            "description": "Carpeta de email"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de emails",
            "content": {
              "application/json": {
                "example": {
                  "status": "ok",
                  "count": 2,
                  "emails": [
                    {
                      "from": "sender@example.com",
                      "to": "receiver@gmail.com",
                      "subject": "Test",
                      "seen": false
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/email/receiveOne": {
      "post": {
        "summary": "Recibir email por ID",
        "description": "Obtiene los detalles de un email específico. **Requiere autenticación**",
        "tags": ["Email"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["messageId"],
                "properties": {
                  "messageId": {
                    "type": "string",
                    "description": "ID del mensaje",
                    "example": "<abc123@example.com>"
                  },
                  "mailbox": {
                    "type": "string",
                    "description": "Carpeta de email",
                    "example": "INBOX"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Detalles del email",
            "content": {
              "application/json": {
                "example": {
                  "status": "ok",
                  "email": {
                    "from": "sender@example.com",
                    "to": "receiver@gmail.com",
                    "subject": "Test",
                    "text": "Contenido",
                    "html": "<p>Contenido</p>",
                    "seen": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Falta messageId"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Email no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/email/receive-sync": {
      "get": {
        "summary": "Sync emails entrantes (El Cartero Programado)",
        "description": "Sincroniza emails desde IMAP hacia la tabla messages. Procesa máximo 5 emails nuevos por ejecución (configurable via EMAIL_SYNC_BATCH_SIZE). Usa checkpoint por messageId para evitar duplicados. **Requiere autenticación**",
        "tags": ["Email"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Sync completado",
            "content": {
              "application/json": {
                "example": {
                  "status": "ok",
                  "processed": 3,
                  "lastProcessedMessageId": "abc123@email.com",
                  "lastSyncAt": "2026-04-05T01:20:00Z"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
    "/api/contacts": {
      "get": {
        "summary": "Listar contactos",
        "description": "Retorna lista paginada de contactos con filtros opcionales. **Requiere autenticación**",
        "tags": ["Contacts"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "number", "default": 1 },
            "description": "Número de página"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "number", "default": 20, "enum": [10, 20, 30, 50, 100] },
            "description": "Registros por página"
          },
          {
            "name": "stage",
            "in": "query",
            "schema": { "type": "string", "enum": ["new", "contacted", "qualified", "proposal", "won", "lost"] },
            "description": "Filtrar por etapa del funnel"
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Buscar en nombre y email"
          },
          {
            "name": "tags",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Tags separados por coma"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de contactos",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "uuid",
                      "name": "Juan Pérez",
                      "email": "juan@example.com",
                      "phone": "+56912345678",
                      "company": "Acme Corp",
                      "stage": "new",
                      "tags": ["vip"],
                      "lastContact": "2026-04-06T10:00:00Z",
                      "notes": null,
                      "avatar": null,
                      "assignedTo": null,
                      "createdAt": "2026-04-06T10:00:00Z",
                      "updatedAt": "2026-04-06T10:00:00Z"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "limit": 20,
                    "total": 50,
                    "totalPages": 3
                  }
                }
              }
            }
          },
          "401": { "description": "No autorizado" }
        }
      },
      "post": {
        "summary": "Crear contacto",
        "description": "Crea un nuevo contacto. **Requiere autenticación**",
        "tags": ["Contacts"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContactCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contacto creado",
            "content": {
              "application/json": {
                "example": {
                  "id": "uuid",
                  "name": "Juan Pérez",
                  "email": "juan@example.com",
                  "phone": "+56912345678",
                  "company": "Acme Corp",
                  "stage": "new",
                  "tags": ["vip"],
                  "notes": null,
                  "assignedTo": null,
                  "lastContact": null,
                  "avatar": null,
                  "createdAt": "2026-04-06T10:00:00Z",
                  "updatedAt": "2026-04-06T10:00:00Z"
                }
              }
            }
          },
          "400": { "description": "Error de validación" },
          "401": { "description": "No autorizado" }
        }
      }
    },
    "/api/contacts/{id}": {
      "get": {
        "summary": "Obtener contacto por ID",
        "description": "Retorna un contacto específico. **Requiere autenticación**",
        "tags": ["Contacts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "ID del contacto"
          }
        ],
        "responses": {
          "200": { "description": "Contacto encontrado" },
          "401": { "description": "No autorizado" },
          "404": { "description": "Contacto no encontrado" }
        }
      },
      "put": {
        "summary": "Actualizar contacto",
        "description": "Actualiza un contacto existente. **Requiere autenticación**",
        "tags": ["Contacts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContactUpdate" }
            }
          }
        },
        "responses": {
          "200": { "description": "Contacto actualizado" },
          "400": { "description": "Error de validación" },
          "401": { "description": "No autorizado" },
          "404": { "description": "Contacto no encontrado" }
        }
      },
      "delete": {
        "summary": "Eliminar contacto",
        "description": "Elimina un contacto. **Requiere autenticación**",
        "tags": ["Contacts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Contacto eliminado" },
          "401": { "description": "No autorizado" },
          "404": { "description": "Contacto no encontrado" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Better Auth session token (cookie: better-auth.session_token)"
      }
    },
...
  },
...
}